Refactor parameters.py run.py
